Moved diversity admin models to root namespace
