+++ app/admin/admin_diversity_responses.rb
@@ -1,4 +1,4 @@
-ActiveAdmin.register Admin::DiversityResponse, as: 'Diversity Responses' do
+ActiveAdmin.register ::DiversityResponse, as: 'Diversity Responses' do
 # See permitted parameters documentation:
 # https://github.com/activeadmin/activeadmin/blob/master/docs/2-resource-customization.md#setting-up-strong-parameters
 #
