--- app/models/admin/diversity_response.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-module Admin
-  class DiversityResponse < ApplicationRecord
-    self.table_name = :diversity_responses
-
-  end
-end
