+++ app/models/diversity_response.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+class DiversityResponse < ApplicationRecord
+  self.table_name = :diversity_responses
+
+end
